+++ src/Console/Stackwatch.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Bakame\Stackwatch\Console;
 
 use Bakame\Stackwatch\Environment;
-use Bakame\Stackwatch\LeaderPrinter;
-use Bakame\Stackwatch\Translator;
+use Bakame\Stackwatch\Exporter\LeaderPrinter;
+use Bakame\Stackwatch\Exporter\Translator;
 use Bakame\Stackwatch\Version;
 use Bakame\Stackwatch\Warning;
 use Psr\Log\LoggerInterface;