+++ src/Exporter/OtlExporter.php
@@ -2,8 +2,19 @@
 
 declare(strict_types=1);
 
-namespace Bakame\Stackwatch;
+namespace Bakame\Stackwatch\Exporter;
 
+use Bakame\Stackwatch\AggregationType;
+use Bakame\Stackwatch\Environment;
+use Bakame\Stackwatch\Metrics;
+use Bakame\Stackwatch\MetricType;
+use Bakame\Stackwatch\Profiler;
+use Bakame\Stackwatch\Report;
+use Bakame\Stackwatch\Result;
+use Bakame\Stackwatch\Snapshot;
+use Bakame\Stackwatch\Span;
+use Bakame\Stackwatch\Statistics;
+use Bakame\Stackwatch\Timeline;
 use OpenTelemetry\API\Metrics\CounterInterface;
 use OpenTelemetry\API\Metrics\HistogramInterface;
 use OpenTelemetry\API\Metrics\MeterInterface;