+++ src/Exporter/StreamWriter.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Bakame\Stackwatch;
+namespace Bakame\Stackwatch\Exporter;
+
+use Bakame\Stackwatch\InvalidArgument;
+use Bakame\Stackwatch\Warning;
 
 use function get_resource_type;
 use function is_resource;